Actualiza productos despues de descarga correcta, control de errores
Productos entrega un texto...

+++ javaServices/InventarioQuarkus/src/main/java/inventario/pica/dominio/InventarioProductoDto.java
@@ -2,7 +2,7 @@ package inventario.pica.dominio;
 
 public class InventarioProductoDto {
 
-    public long id;
-    public long inventario;
+    public int id;
+    public int inventario;
 
 }
